Change all occurences of 'video' to 'lesson' in LessonButtonCard storybook file
